+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 98f68b5e6a8399c4ea79884323df584e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
