+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Bounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 985f324ad0d5f654892438f04ff6c91a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
