+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Centered.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a70e83e5ce627a640b94ada2e0fc0df6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
