+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/IMinimap.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: efa885e7ad150a54abbf1ad7eb5ee107
+guid: c68e8cd819ce4724b98d5ea2f5afeff0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
