+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/MinimapBounds.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 704283e4901c96444b85bd5de059b89d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
