+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/MinimapBounds/BoxBoundsSource.cs
@@ -16,35 +16,21 @@
 // ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-using com.nickmaltbie.MinimapTools.Icon;
 using UnityEngine;
 
-namespace com.nickmaltbie.MinimapTools
+namespace com.nickmaltbie.MinimapTools.Minimap.MinimapBounds
 {
     /// <summary>
-    /// Minimap for managing state.
+    /// Source of bounds from a box collider
     /// </summary>
-    public interface IMinimap
+    [RequireComponent(typeof(BoxCollider))]
+    public class BoxBoundsSource : MonoBehaviour, IBoundsSource
     {
-        /// <summary>
-        /// Check if a location is in the minimap space.
-        /// </summary>
-        /// <param name="worldSpace">World space of the object.</param>
-        /// <returns>True if the location is within the minimap, false otherwise.</returns>
-        bool InMap(Vector3 worldSpace);
-
-        /// <summary>
-        /// Add an icon to the minimap.
-        /// </summary>
-        /// <param name="minimapIcon">Icon to add to the minimap.</param>
-        /// <returns>True if the object was added successfully, false otherwise.</returns>
-        bool AddIcon(IMinimapIcon minimapIcon);
-
-        /// <summary>
-        /// Remove an icon from the minimap.
-        /// </summary>
-        /// <param name="minimapIcon">Icon to remove to the minimap.</param>
-        /// <returns>True if the object was removed successfully, false otherwise.</returns>
-        bool RemoveIcon(IMinimapIcon minimapIcon);
+        /// <inheritdoc/>
+        public Bounds GetBounds()
+        {
+            Bounds bounds = GetComponent<BoxCollider>().bounds;
+            return new Bounds(bounds.center + transform.position, bounds.size);
+        }
     }
 }
