+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/MinimapBounds/BoxBoundsSource.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5528c4a086ef5a243aeabf895f7c5b76
+guid: 3e951840b681a4b4b99a068647664edb
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
