+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/MinimapBounds/IBoundsSource.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3837591cdf4f54040abcdc22792b5d40
+guid: 4cefdae7ff9719a4fa3507e48cad714b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
