+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Simple.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 61e8948bb375aeb4c842f2927342d350
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
