+++ Packages/com.nickmaltbie.MinimapTools/Scripts/Minimap/Simple/SimpleStaticMinimap.cs.meta
